refactor cells to protect their IDs
