+++ base/src/main/java/com/smartdevicelink/managers/screen/BaseScreenManager.java
@@ -40,8 +40,8 @@ import com.smartdevicelink.managers.file.FileManager;
 import com.smartdevicelink.managers.file.filetypes.SdlArtwork;
 import com.smartdevicelink.managers.screen.menu.MenuManager;
 import com.smartdevicelink.managers.screen.menu.VoiceCommandManager;
-import com.smartdevicelink.managers.screen.menu.cells.MenuCell;
-import com.smartdevicelink.managers.screen.menu.cells.VoiceCommand;
+import com.smartdevicelink.managers.screen.menu.MenuCell;
+import com.smartdevicelink.managers.screen.menu.VoiceCommand;
 import com.smartdevicelink.proxy.interfaces.ISdl;
 import com.smartdevicelink.proxy.rpc.enums.MetadataType;
 import com.smartdevicelink.proxy.rpc.enums.TextAlignment;
